requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
nose==1.3.7
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
virtualenv==16.6.1
whitenoise==4.1.3
